requirements.txt
asgiref==3.7.2
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
reportlab==4.0.6
setuptools==68.2.2
sqlparse==0.4.4
tzdata==2023.3